Исправлена гора варнингов.

+++ tutorial30/camera.cpp
@@ -43,26 +43,26 @@ void Camera::Init()
     {
         if (HTarget.x >= 0.0f)
         {
-            m_AngleH = 360.0f - ToDegree(asin(HTarget.z));
+            m_AngleH = 360.0f - ToDegree(asinf(HTarget.z));
         }
         else
         {
-            m_AngleH = 180.0f + ToDegree(asin(HTarget.z));
+            m_AngleH = 180.0f + ToDegree(asinf(HTarget.z));
         }
     }
     else
     {
         if (HTarget.x >= 0.0f)
         {
-            m_AngleH = ToDegree(asin(-HTarget.z));
+            m_AngleH = ToDegree(asinf(-HTarget.z));
         }
         else
         {
-            m_AngleH = 90.0f + ToDegree(asin(-HTarget.z));
+            m_AngleH = 90.0f + ToDegree(asinf(-HTarget.z));
         }
     }
 
-    m_AngleV = -ToDegree(asin(m_target.y));
+    m_AngleV = -ToDegree(asinf(m_target.y));
 
     m_mousePos.x  = m_windowWidth / 2;
     m_mousePos.y  = m_windowHeight / 2;
